LPS-87113 Use activator to initialize Audit service release number
