+++ modules/apps/portal-security-audit/portal-security-audit-storage-service/bnd.bnd
@@ -1,3 +1,4 @@
+Bundle-Activator: com.liferay.portal.security.audit.storage.internal.activator.AuditStorageServiceBundleActivator
 Bundle-Name: Liferay Portal Security Audit Storage Service
 Bundle-SymbolicName: com.liferay.portal.security.audit.storage.service
 Bundle-Version: 4.0.0
