+++ modules/apps/portal-security-audit/portal-security-audit-storage-service/src/main/java/com/liferay/portal/security/audit/storage/internal/upgrade/AuditStorageServiceUpgrade.java
@@ -15,9 +15,7 @@
 package com.liferay.portal.security.audit.storage.internal.upgrade;
 
 import com.liferay.portal.kernel.upgrade.DummyUpgradeStep;
-import com.liferay.portal.kernel.upgrade.UpgradeException;
 import com.liferay.portal.upgrade.registry.UpgradeStepRegistrator;
-import com.liferay.portal.upgrade.release.BaseUpgradeServiceModuleRelease;
 
 import org.osgi.service.component.annotations.Component;
 
@@ -29,34 +27,6 @@ public class AuditStorageServiceUpgrade implements UpgradeStepRegistrator {
 
 	@Override
 	public void register(Registry registry) {
-		try {
-			BaseUpgradeServiceModuleRelease upgradeServiceModuleRelease =
-				new BaseUpgradeServiceModuleRelease() {
-
-					@Override
-					protected String getNamespace() {
-						return "Audit";
-					}
-
-					@Override
-					protected String getNewBundleSymbolicName() {
-						return
-							"com.liferay.portal.security.audit.storage.service";
-					}
-
-					@Override
-					protected String getOldBundleSymbolicName() {
-						return "audit-portlet";
-					}
-
-				};
-
-			upgradeServiceModuleRelease.upgrade();
-		}
-		catch (UpgradeException ue) {
-			throw new RuntimeException(ue);
-		}
-
 		registry.register(
 			"com.liferay.portal.security.audit.storage.service", "0.0.1",
 			"1.0.0", new DummyUpgradeStep());
